pyproject.toml
[project]
name = "serbia-dish-image-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.10.1",
    "bs4>=0.0.2",
    "diffusers>=0.34.0,<0.35",
    "dotenv>=0.9.9",
    "numpy<2",
    "open-clip-torch>=3.2.0",
    "openai>=2.1.0",
    "peft>=0.17.1",
    "pillow>=11.3.0",
    "requests>=2.28.1",
    "safetensors>=0.6.2",
    "selenium>=4.36.0",
    "torch>=2.3,<2.6",
    "torchvision>=0.18,<0.19",
    "torchaudio>=2.3,<2.6",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "xformers>=0.0.20"
]

[tool.uv.sources]
torch = { index = "torch-cu121" }
torchvision = { index = "torch-cu121" }
torchaudio = { index = "torch-cu121" }
xformers = { index = "torch-cu121" }

[[tool.uv.index]]
name = "torch-cu121"
url = "https://download.pytorch.org/whl/cu121"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true
